+++ eras/byron/ledger/impl/src/Cardano/Chain/Update/SystemTag.hs
@@ -12,8 +12,6 @@ module Cardano.Chain.Update.SystemTag (
   SystemTagError (..),
   checkSystemTag,
   systemTagMaxLength,
-  osHelper,
-  archHelper,
 )
 where
 
@@ -36,8 +34,6 @@ import Cardano.Prelude hiding (cborError)
 import Data.Aeson (ToJSON, ToJSONKey)
 import Data.Data (Data)
 import qualified Data.Text as T
-import Distribution.System (Arch (..), OS (..))
-import Distribution.Text (display)
 import Formatting (bprint, int, stext)
 import qualified Formatting.Buildable as B
 import NoThunks.Class (NoThunks (..))
@@ -119,20 +115,3 @@ checkSystemTag (SystemTag tag)
   | T.length tag > systemTagMaxLength = throwError $ SystemTagTooLong tag
   | T.any (not . isAscii) tag = throwError $ SystemTagNotAscii tag
   | otherwise = pure ()
-
--- | Helper to turn an @OS@ into a @Text@ compatible with the @systemTag@
---   previously used in 'configuration.yaml'
-osHelper :: OS -> Text
-osHelper sys = case sys of
-  Windows -> "win"
-  OSX -> "macos"
-  Linux -> "linux"
-  _ -> toS $ display sys
-
--- | Helper to turn an @Arch@ into a @Text@ compatible with the @systemTag@
---   previously used in 'configuration.yaml'
-archHelper :: Arch -> Text
-archHelper archt = case archt of
-  I386 -> "32"
-  X86_64 -> "64"
-  _ -> toS $ display archt