+++ eras/byron/ledger/impl/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for `cardano-ledger-byron`
 
+## 1.3.0.0
+
+* Remove `osHelper` and `archHelper` functions.
+
 ## 1.2.0.0
 
 * Added `Test.Cardano.Chain.Binary.Cddl` module